In the search of a good RNG

+++ NotThatGameEngine/NotThatGameEngine/Component.cpp
@@ -3,7 +3,7 @@
 #include "Application.h"
 
 
-Component::Component(int _id, GameObject* _gameObject, COMPONENT_TYPE _type, bool _enabled) : type(_type), enabled(_enabled), owner(_gameObject), id(_id), deleteComponent(false) {}
+Component::Component(long long int _id, GameObject* _gameObject, COMPONENT_TYPE _type, bool _enabled) : type(_type), enabled(_enabled), owner(_gameObject), id(_id), deleteComponent(false) {}
 
 
 Component::~Component() { owner = nullptr; }
