+++ NotThatGameEngine/NotThatGameEngine/GameObject.h
@@ -11,15 +11,15 @@ class GameObject {
 
 public:
 
-	GameObject(int _id, std::string _name = "NewGameObject", GameObject* parent = nullptr, bool enabled = true, std::vector<GameObject*> children = std::vector<GameObject*>());
+	GameObject(long long int _id, std::string _name = "NewGameObject", GameObject* parent = nullptr, bool enabled = true, std::vector<GameObject*> children = std::vector<GameObject*>());
 	~GameObject();
 
 	void Update();
 	void PostUpdate(uint& defaultTextureId);
 	Component* AddComponent(COMPONENT_TYPE _type);
 	bool AddGameObjectByParent(GameObject* newObject);
-	int GenerateComponentId();
-	bool CheckChildDeletionById(int _id);
+	long long int GenerateComponentId();
+	bool CheckChildDeletionById(long long int _id);
 	void SetDeleteGameObject();
 	Component* FindComponent(COMPONENT_TYPE _type);
 	std::vector<Component*> FindComponents(COMPONENT_TYPE _type);
@@ -31,7 +31,7 @@ public:
 	float4x4 worldTransform;
 	std::vector<GameObject*> childs;
 
-	int id;
+	long long int id;
 	bool enabled;
 	std::vector<Component*> components;
