+++ NotThatGameEngine/NotThatGameEngine/Material.cpp
@@ -1,6 +1,6 @@
 #include "Material.h"
 
-Material::Material(int _id, GameObject* _gameObject) : Component(_id, _gameObject, COMPONENT_TYPE::MATERIAL), diffuseId(0) {}
+Material::Material(long long int _id, GameObject* _gameObject) : Component(_id, _gameObject, COMPONENT_TYPE::MATERIAL), diffuseId(0) {}
 
 
 Material::~Material() {}
