+++ NotThatGameEngine/NotThatGameEngine/Mesh.cpp
@@ -1,6 +1,6 @@
 #include "Mesh.h"
 
-Mesh::Mesh(int _id, GameObject* _gameObject) : Component(_id, _gameObject, COMPONENT_TYPE::MESH), vertexId(0), indexId(0), normalsId(0), textureCoordId(0),
+Mesh::Mesh(long long int _id, GameObject* _gameObject) : Component(_id, _gameObject, COMPONENT_TYPE::MESH), vertexId(0), indexId(0), normalsId(0), textureCoordId(0),
 vertices(), normals(), textureCoord(), indices(), meshName("No_name_Mesh"), paintNormals(false), debugNormalsId(0), debugNormals() {}
